reorganized files, created provider component, cleaned up example
