+++ examples/basic/Root.tsx
@@ -1,0 +1,13 @@
+import * as React from "react"
+import { DevToolkitProvider } from "@TMMF/react-dev-toolkit"
+import Component from "./Component"
+
+const Root = () => {
+  return (
+    <DevToolkitProvider>
+      <Component />
+    </DevToolkitProvider>
+  )
+}
+
+export default Root
